standardizzato l'interfaccia per le tre risorse e spostata la logica di mapping da model a dto in un unico file (app.mapper) per evitare dipendenze circolari di classi

+++ src/utente/utente.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { Prisma, Utente } from '../generated/prisma/client.js';
+import { Prisma } from '../generated/prisma/client.js';
 import { PrismaService } from '../prisma.service.js';
 import { CreateUtenteDto } from './dto/create-utente.dto.js';
 import { UpdateUtenteDto } from './dto/update-utente.dto.js';
@@ -7,20 +7,25 @@ import { UtenteSelectParamsDto } from './dto/utente-select-params.dto.js';
 import { ResponseUtenteDto } from './dto/response-utente.dto.js';
 import { UtenteMapper } from './utente.mapper.js';
 import { UtenteFindAllParamsDto } from './utente.controller.js';
+import { DeepPartial } from './utente.mapper.js';
+import { utenteSelect } from './utente.select.js';
+import { AppMapper } from '../app.mapper.js';
 
 @Injectable()
 export class UtenteService {
   constructor(
     private prisma: PrismaService,
     private mapper: UtenteMapper,
+    private appMapper: AppMapper,
   ) {}
 
   async create(data: CreateUtenteDto): Promise<ResponseUtenteDto> {
     try {
       const utente = await this.prisma.utente.create({
         data: this.mapper.createDtoToModel(data),
+        select: utenteSelect,
       });
-      return this.mapper.modelToDto(utente);
+      return this.appMapper.mapUtente(utente) as ResponseUtenteDto;
     } catch (error) {
       if (
         error instanceof Prisma.PrismaClientKnownRequestError &&
@@ -37,17 +42,17 @@ export class UtenteService {
 
   async findAll(
     params: UtenteFindAllParamsDto,
-  ): Promise<Partial<ResponseUtenteDto>[]> {
-    const utenti: Partial<Utente>[] = await this.prisma.utente.findMany(
+  ): Promise<DeepPartial<ResponseUtenteDto>[]> {
+    const utenti = await this.prisma.utente.findMany(
       this.mapper.allParamsDtoToModel(params),
     );
-    return utenti.map((utente) => this.mapper.modelToPartialDto(utente));
+    return utenti.map((utente) => this.appMapper.mapUtente(utente));
   }
 
   async findOne(
     id: number,
     params: UtenteSelectParamsDto,
-  ): Promise<Partial<ResponseUtenteDto>> {
+  ): Promise<DeepPartial<ResponseUtenteDto>> {
     const utente = await this.prisma.utente.findUnique({
       where: { id },
       select: this.mapper.selectParamsDtoToModel(params),
@@ -55,7 +60,7 @@ export class UtenteService {
     if (!utente) {
       throw new NotFoundException(`Utente ${id} not found`);
     }
-    return this.mapper.modelToPartialDto(utente);
+    return this.appMapper.mapUtente(utente);
   }
 
   async update(
@@ -66,9 +71,10 @@ export class UtenteService {
       const utente = await this.prisma.utente.update({
         where: { id },
         data: this.mapper.updateDtoToModel(update),
+        select: utenteSelect,
       });
 
-      return this.mapper.modelToDto(utente);
+      return this.appMapper.mapUtente(utente) as ResponseUtenteDto;
     } catch (error) {
       if (
         error instanceof Prisma.PrismaClientKnownRequestError &&
@@ -83,10 +89,16 @@ export class UtenteService {
 
   async remove(id: number): Promise<ResponseUtenteDto> {
     try {
-      const utente = await this.prisma.utente.delete({
+      const utente = await this.prisma.utente.findUnique({ where: { id } });
+      if (!utente) {
+        throw new NotFoundException(`Utente ${id} not found`);
+      }
+
+      const deletedUtente = await this.prisma.utente.delete({
         where: { id },
+        select: utenteSelect,
       });
-      return this.mapper.modelToDto(utente);
+      return this.appMapper.mapUtente(deletedUtente) as ResponseUtenteDto;
     } catch (error) {
       if (
         error instanceof Prisma.PrismaClientKnownRequestError &&
